Extract MetricsPushOptions and refactor away MetricsEndpoint
