+++ commandline/src/main/java/tech/pegasys/web3signer/commandline/Web3SignerBaseCommand.java
@@ -12,20 +12,20 @@
  */
 package tech.pegasys.web3signer.commandline;
 
-import static org.hyperledger.besu.metrics.prometheus.MetricsConfiguration.DEFAULT_METRICS_PUSH_PORT;
 import static tech.pegasys.web3signer.commandline.DefaultCommandValues.CONFIG_FILE_OPTION_NAME;
 import static tech.pegasys.web3signer.commandline.DefaultCommandValues.FILE_FORMAT_HELP;
 import static tech.pegasys.web3signer.commandline.DefaultCommandValues.HOST_FORMAT_HELP;
-import static tech.pegasys.web3signer.commandline.DefaultCommandValues.INTEGER_FORMAT_HELP;
 import static tech.pegasys.web3signer.commandline.DefaultCommandValues.PORT_FORMAT_HELP;
 import static tech.pegasys.web3signer.common.Web3SignerMetricCategory.DEFAULT_METRIC_CATEGORIES;
 
 import tech.pegasys.web3signer.commandline.config.AllowListHostsProperty;
+import tech.pegasys.web3signer.commandline.config.PicoCliMetricsPushOptions;
 import tech.pegasys.web3signer.commandline.config.PicoCliTlsServerOptions;
 import tech.pegasys.web3signer.commandline.config.PicoCliTlsServerOptionsValidator;
 import tech.pegasys.web3signer.commandline.convertor.MetricCategoryConverter;
 import tech.pegasys.web3signer.common.Web3SignerMetricCategory;
 import tech.pegasys.web3signer.core.config.BaseConfig;
+import tech.pegasys.web3signer.core.config.MetricsPushOptions;
 import tech.pegasys.web3signer.core.config.TlsOptions;
 
 import java.io.File;
@@ -175,40 +175,7 @@ public class Web3SignerBaseCommand implements BaseConfig, Runnable {
       defaultValue = "localhost,127.0.0.1")
   private final AllowListHostsProperty metricsHostAllowList = new AllowListHostsProperty();
 
-  @Option(
-      names = {"--metrics-push-enabled"},
-      description = "Enable the metrics push gateway integration (default: ${DEFAULT-VALUE})")
-  private final Boolean isMetricsPushEnabled = false;
-
-  @SuppressWarnings({"FieldCanBeFinal", "FieldMayBeFinal"}) // PicoCLI requires non-final Strings.
-  @Option(
-      names = {"--metrics-push-host"},
-      paramLabel = HOST_FORMAT_HELP,
-      description = "Host of the Prometheus Push Gateway for push mode (default: ${DEFAULT-VALUE})",
-      arity = "1")
-  private String metricsPushHost = InetAddress.getLoopbackAddress().getHostAddress();
-
-  @Option(
-      names = {"--metrics-push-port"},
-      paramLabel = PORT_FORMAT_HELP,
-      description = "Port of the Prometheus Push Gateway for push mode (default: ${DEFAULT-VALUE})",
-      arity = "1")
-  private final Integer metricsPushPort = DEFAULT_METRICS_PUSH_PORT;
-
-  @Option(
-      names = {"--metrics-push-interval"},
-      paramLabel = INTEGER_FORMAT_HELP,
-      description =
-          "Interval in seconds to push metrics when in push mode (default: ${DEFAULT-VALUE})",
-      arity = "1")
-  private final Integer metricsPushIntervalSeconds = 15;
-
-  @SuppressWarnings({"FieldCanBeFinal", "FieldMayBeFinal"}) // PicoCLI requires non-final Strings.
-  @Option(
-      names = {"--metrics-push-prometheus-job"},
-      description = "Job name to use when in push mode (default: ${DEFAULT-VALUE})",
-      arity = "1")
-  private String metricsPrometheusJob = "web3signer-job";
+  @CommandLine.Mixin private PicoCliMetricsPushOptions metricsPushOptions;
 
   @Option(
       names = {"--idle-connection-timeout-seconds"},
@@ -304,28 +271,11 @@ public class Web3SignerBaseCommand implements BaseConfig, Runnable {
   }
 
   @Override
-  public Boolean isMetricsPushEnabled() {
-    return isMetricsPushEnabled;
-  }
-
-  @Override
-  public String getMetricsPushHost() {
-    return metricsPushHost;
-  }
-
-  @Override
-  public Integer getMetricsPushPort() {
-    return metricsPushPort;
-  }
-
-  @Override
-  public Integer getMetricsPushIntervalSeconds() {
-    return metricsPushIntervalSeconds;
-  }
-
-  @Override
-  public String getMetricsPrometheusJob() {
-    return metricsPrometheusJob;
+  public Optional<MetricsPushOptions> getMetricsPushOptions() {
+    if (metricsPushOptions.isMetricsPushEnabled()) {
+      return Optional.of(metricsPushOptions);
+    }
+    return Optional.empty();
   }
 
   @Override
@@ -397,7 +347,7 @@ public class Web3SignerBaseCommand implements BaseConfig, Runnable {
         new PicoCliTlsServerOptionsValidator(spec, picoCliTlsServerOptions);
     picoCliTlsServerOptionsValidator.validate();
 
-    if (isMetricsEnabled() && isMetricsPushEnabled()) {
+    if (isMetricsEnabled() && metricsPushOptions.isMetricsPushEnabled()) {
       throw new CommandLine.MutuallyExclusiveArgsException(
           spec.commandLine(),
           "--metrics-enabled option and --metrics-push-enabled option can't be used at the same "
